version 2.0/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
lxml==5.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
mysql-connector-python==8.3.0
numpy==1.26.3
pandas==2.2.0
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.4
requests==2.31.0
rich==13.7.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.27.1
wheel==0.41.2
